frontend package updates: react and redux
