+++ packages/frontend-web/package.json
@@ -82,15 +82,15 @@
     "@types/prop-types": "^15.7.3",
     "@types/qs": "^6.9.5",
     "@types/randomstring": "^1.1.6",
-    "@types/react": "16.9.15",
-    "@types/react-custom-scrollbars": "4.0.6",
-    "@types/react-dom": "16.9.4",
-    "@types/react-linkify": "1.0.0",
-    "@types/react-redux": "7.1.3",
-    "@types/react-router-dom": "5.1.3",
-    "@types/react-virtualized-auto-sizer": "1.0.0",
-    "@types/react-window": "1.8.2",
-    "@types/redux-actions": "2.6.1",
+    "@types/react": "^17.0.0",
+    "@types/react-custom-scrollbars": "^4.0.7",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-linkify": "^1.0.0",
+    "@types/react-redux": "^7.1.14",
+    "@types/react-router-dom": "^5.1.6",
+    "@types/react-virtualized-auto-sizer": "^1.0.0",
+    "@types/react-window": "^1.8.2",
+    "@types/redux-actions": "^2.6.1",
     "aphrodite": "2.3.1",
     "axios": "^0.21.1",
     "check-types": "^11.1.2",
@@ -99,7 +99,7 @@
     "date-fns": "^2.16.1",
     "enzyme": "^3.11.0",
     "faker": "4.1.0",
-    "focus-trap-react": "6.0.0",
+    "focus-trap-react": "^8.3.2",
     "i": "^0.3.6",
     "immutable": "3.8.2",
     "jwt-decode": "^3.1.2",
@@ -111,22 +111,22 @@
     "qs": "^6.9.4",
     "query-string": "^6.13.8",
     "randomstring": "^1.1.5",
-    "react": "^16.14.0",
-    "react-a11y": "1.1.0",
-    "react-custom-scrollbars": "4.2.1",
-    "react-dom": "16.12.0",
-    "react-draggable": "4.2.0",
+    "react": "^17.0.1",
+    "react-a11y": "^1.1.0",
+    "react-custom-scrollbars": "^4.2.1",
+    "react-dom": "^17.0.1",
+    "react-draggable": "^4.4.3",
     "react-linkify": "1.0.0-alpha",
-    "react-perfect-scrollbar": "1.5.3",
-    "react-redux": "7.1.3",
-    "react-router-dom": "5.1.2",
-    "react-virtualized-auto-sizer": "1.0.2",
-    "react-window": "1.8.5",
-    "redux": "4.0.4",
-    "redux-actions": "2.6.5",
+    "react-perfect-scrollbar": "^1.5.8",
+    "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
+    "react-virtualized-auto-sizer": "^1.0.3",
+    "react-window": "^1.8.6",
+    "redux": "^4.0.5",
+    "redux-actions": "^2.6.5",
     "redux-devtools": "^3.7.0",
-    "redux-thunk": "2.3.0",
-    "reselect": "4.0.0",
+    "redux-thunk": "^2.3.0",
+    "reselect": "^4.0.0",
     "slugify": "^1.4.6",
     "typescript": "^4.1.3"
   }
